requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pydantic==2.6.1
jinja2==3.1.3
pyyaml==6.0.1
python-dotenv==1.0.1
